--- src/main/java/umc/spring/service/MemberService/MemberCommandAndServiceImpl.java
@@ -1,4 +1,0 @@
-package umc.spring.service.MemberService;
-
-public class MemberCommandAndServiceImpl {
-}
