+++ src/main/java/umc/spring/service/MemberService/MemberCommandServiceImpl.java
@@ -1,0 +1,20 @@
+package umc.spring.service.MemberService;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import umc.spring.domain.mapping.Member;
+import umc.spring.repository.MemberRepository;
+import umc.spring.web.dto.MemberRequestDTO;
+
+@Service
+@RequiredArgsConstructor
+public class MemberCommandServiceImpl implements MemberCommandService{
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public Member joinMember(MemberRequestDTO.JoinDto request) {
+
+        return null;
+    }
+}
